+++ mirrord/config/src/lib.rs
@@ -126,6 +126,13 @@ pub struct LayerConfig {
     /// in-cluster config.
     #[config(env = "MIRRORD_KUBECONFIG")]
     pub kubeconfig: Option<String>,
+
+    /// Binaries to patch (macOS SIP).
+    /// Use this when mirrord isn't loaded to protected binaries
+    /// that weren't automatically patched.
+    /// Runs `endswith` on the binary path (so `bash` would apply to any binary
+    /// ending with `bash` while `/usr/bin/bash` would apply only for that binary).
+    pub sip_binaries: Option<VecOrSingle<String>>,
 }
 
 impl LayerConfig {
@@ -377,6 +384,7 @@ mod tests {
             }),
             connect_tcp: None,
             operator: None,
+            sip_binaries: None,
         };
 
         assert_eq!(config, expect);